Remove out of date binary send example and add minimal replacement
The provided python example was out of date. Writing a new one seems futile though as the API is likely to change. Instead this commit adds links to the relevant serialization code so people can easily reimplement it. The links are to the current HEAD of the develop branch, as there is no reliable way to link to a semantic element instead of a line, which might change later on.