new first section for newcomers to nym
A lot of people(newcomers to Nym) have been asking me stuff such as:
- where are the binaries after I build Nym?
- how do I get the binaries or build them? 

If my memory serves me right, this section used to be here in the past, although without the exact mixnode binary path. 

I think it is a proper addition for the docs, even if the person has some IT/dev knowledge, they might be compiling a Rust program for the first time and they won't know where the binaries can be unless they really search for them.